document_label_service/requirements.txt
fastapi
uvicorn
requests
python-dotenv
sqlalchemy
psycopg2-binary
loguru
